Rename cpu_solver to openmp_solver
