+++ .github/workflows/sonarqube.yml
@@ -34,13 +34,13 @@ jobs:
           LLVM_PROFILE_FILE="common.profraw" build/llvm-cov-debug/common/physicore.common.tests
           LLVM_PROFILE_FILE="biofvm.profraw" build/llvm-cov-debug/reactions-diffusion/biofvm/physicore.reactions-diffusion.biofvm.tests
           LLVM_PROFILE_FILE="micromechanics.profraw" build/llvm-cov-debug/mechanics/physicell/physicore.mechanics.physicell.tests
-          LLVM_PROFILE_FILE="biofvm.cpu_solver.profraw" build/llvm-cov-debug/reactions-diffusion/biofvm/kernels/cpu_solver/tests/physicore.reactions-diffusion.kernels.cpu_solver.tests
+          LLVM_PROFILE_FILE="biofvm.openmp_solver.profraw" build/llvm-cov-debug/reactions-diffusion/biofvm/kernels/openmp_solver/tests/physicore.reactions-diffusion.kernels.openmp_solver.tests
       - name: Merge counters from different runs and convert the result into an llvm-cov-compatible form
         run: |
-          llvm-profdata-18 merge -o merged.profdata common.profraw biofvm.profraw micromechanics.profraw biofvm.cpu_solver.profraw
+          llvm-profdata-18 merge -o merged.profdata common.profraw biofvm.profraw micromechanics.profraw biofvm.openmp_solver.profraw
       - name: Collect the coverage
         run: |
-          llvm-cov-18 show --show-branches=count --instr-profile merged.profdata build/llvm-cov-debug/common/physicore.common.tests -object build/llvm-cov-debug/reactions-diffusion/biofvm/physicore.reactions-diffusion.biofvm.tests -object build/llvm-cov-debug/mechanics/physicell/physicore.mechanics.physicell.tests -object build/llvm-cov-debug/reactions-diffusion/biofvm/kernels/cpu_solver/tests/physicore.reactions-diffusion.kernels.cpu_solver.tests > coverage.txt
+          llvm-cov-18 show --show-branches=count --instr-profile merged.profdata build/llvm-cov-debug/common/physicore.common.tests -object build/llvm-cov-debug/reactions-diffusion/biofvm/physicore.reactions-diffusion.biofvm.tests -object build/llvm-cov-debug/mechanics/physicell/physicore.mechanics.physicell.tests -object build/llvm-cov-debug/reactions-diffusion/biofvm/kernels/openmp_solver/tests/physicore.reactions-diffusion.kernels.openmp_solver.tests > coverage.txt
       - name: SonarQube Scan
         uses: SonarSource/sonarqube-scan-action@v4
         env:
@@ -51,6 +51,6 @@ jobs:
             --define sonar.organization=bsc-life
             --define sonar.cfamily.compile-commands="${{ env.BUILD_WRAPPER_OUT_DIR }}/compile_commands.json"
             --define sonar.cfamily.llvm-cov.reportPath=coverage.txt
-            --define sonar.sources=phenotype/physicore/src,common/src,common/include,reactions-diffusion/biofvm/src,reactions-diffusion/biofvm/include,mechanics/physicell/src,mechanics/physicell/include,reactions-diffusion/biofvm/kernels/cpu_solver/src,reactions-diffusion/biofvm/kernels/cpu_solver/examples,reactions-diffusion/biofvm/kernels/cpu_solver/include
-            --define sonar.tests=common/tests,reactions-diffusion/biofvm/tests,mechanics/physicell/tests,reactions-diffusion/biofvm/kernels/cpu_solver/tests
-            --define sonar.coverage.exclusions=reactions-diffusion/biofvm/kernels/cpu_solver/examples/*
+            --define sonar.sources=phenotype/physicore/src,common/src,common/include,reactions-diffusion/biofvm/src,reactions-diffusion/biofvm/include,mechanics/physicell/src,mechanics/physicell/include,reactions-diffusion/biofvm/kernels/openmp_solver/src,reactions-diffusion/biofvm/kernels/openmp_solver/examples,reactions-diffusion/biofvm/kernels/openmp_solver/include
+            --define sonar.tests=common/tests,reactions-diffusion/biofvm/tests,mechanics/physicell/tests,reactions-diffusion/biofvm/kernels/openmp_solver/tests
+            --define sonar.coverage.exclusions=reactions-diffusion/biofvm/kernels/openmp_solver/examples/*
