+++ reactions-diffusion/biofvm/kernels/openmp_solver/src/bulk_solver.cpp
